back/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
tqdm
matplotlib
seaborn